App keep crashing when trying to create default Adapter, save before making custom Adapter
