+++ app/src/main/res/layout/activity_add_course.xml
@@ -12,8 +12,8 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="0dp"
-        android:layout_weight="0.18"
-        android:text="Enter a course code or course name, then click on the 'Add' button to add a course" />
+        android:layout_weight="0.21"
+        android:text="Enter a course key, then click on the 'Add' button to add a course" />
 
     <EditText
         android:id="@+id/enter_course"
@@ -25,9 +25,20 @@
         android:id="@+id/add_course_button"
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="100dip"
+        android:layout_marginTop="10dip"
+        android:layout_marginBottom="20dip"
         android:background="@color/colorAccent"
         android:text="Add course"
         android:textColor="@android:color/black" />
+
+    <ListView
+        android:id="@+id/courseview"
+        android:layout_width="match_parent"
+        android:layout_height="204dp"
+        android:background="@null"
+        android:divider="@null"
+        android:stackFromBottom="true"
+        android:transcriptMode="alwaysScroll"
+        android:layout_weight="0.35" />
 
 </LinearLayout>
