--- app/src/main/res/layout/activity_choose_course.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".course.ChooseCourseActivity"
-    android:weightSum="1">
-
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_weight="0.16"
-        android:text="Choose a course by clicking on it" />
-
-    <ListView
-        android:id="@+id/courseview"
-        android:layout_width="fill_parent"
-        android:layout_height="283dp"
-        android:layout_weight="0.75"
-        android:background="@null"
-        android:divider="@null"
-        android:stackFromBottom="true"
-        android:transcriptMode="alwaysScroll" />
-
-</LinearLayout>
